Add compression and binary file expansion

+++ src/main/kotlin/com/github/dimguilpatcher/patcher/PatcherRule.kt
@@ -3,7 +3,7 @@ package com.github.dimguilpatcher.patcher
 import com.github.dimguilpatcher.json.PatcherRuleSerializer
 import kotlinx.serialization.Serializable
 
-typealias Address = UInt
+typealias Address = ULong
 
 @Serializable(with = PatcherRuleSerializer::class)
 data class PatcherRule(val file: String, val edits: Map<Address, List<Byte>>)
